backend/requirements.txt.txt
Flask
flask-cors
requests
pyfiglet
colorama
pyserial
indic-transliteration